// frontend/jobapptrackerfrontend/src/enums/jobappcardtestids/JobAppCardTestIds_enum.ts
export enum JobAppCardTestIds {
    jobTitle = "JobAppCardType",
    company = "JobAppCardCompany",
    status = "JobAppCardStatus",
    dateApplied = "JobAppCardDateApplied",
    editButton = "JobAppCardEditButton",
    deleteButton = "JobAppCardDeleteButton",
    viewButton = "JobAppCardViewButton"
};