// frontend/jobapptrackerfrontend/src/enums/DeleteJobAppTestIds_enum.ts
export enum DeleteJobAppTestIds {
    closeButton = "DeleteJobAppAlert_close",
    alertHeader = "DeleteJobAppAlert_AlertHeader",
    alert = "DeleteJobAppAlert_Alert",
    alertBody = "DeleteJobAppAlert_AlertBody",
    deleteButton = "DeleteJobAppAlert_DeleteButton",
    cancelButton = "DeleteJobAppAlert_CancelButton"
};